requirements.txt
fastapi[standard]
pymongo
python-dotenv
python-multipart
cloudinary
UVicorn[standard]
requests
fastapi [all]